searcing robot library in /usr/lib/robots/<robot library name> folder
Before it was /usr/lib/robots. This is because now the glue is
installing also the urdf file associated to the robot in the same folder
where there is the shared library.
So having subfolders in /usr/lib/robots for each library makes the file
organization cleaner. In /usr/lib/robots/aliengolib wìyou have both
libaliengolib.so and aliengo.urdf